+++ rig/rig-core/src/providers/internal/mod.rs
@@ -1,7 +1,7 @@
 //! Internal provider helpers shared across provider implementations.
 
 pub(crate) mod buffered;
-pub(crate) mod chat_compatible;
+pub(crate) mod openai_chat_completions_compatible;
 
 pub(crate) fn completion_usage(
     input_tokens: u64,
