+++ rig/rig-core/src/providers/openrouter/streaming.rs
@@ -6,7 +6,7 @@ use tracing::info_span;
 use crate::completion::{CompletionError, CompletionRequest, GetTokenUsage};
 use crate::http_client::HttpClientExt;
 use crate::json_utils;
-use crate::providers::internal::chat_compatible::{
+use crate::providers::internal::openai_chat_completions_compatible::{
     self, CompatibleChoiceData, CompatibleChunk, CompatibleFinishReason, CompatibleStreamProfile,
     CompatibleToolCallChunk,
 };
@@ -202,23 +202,27 @@ impl CompatibleStreamProfile for OpenRouterCompatibleProfile {
             }
         };
 
-        Ok(Some(chat_compatible::normalize_first_choice_chunk(
-            Some(data.id),
-            Some(data.model),
-            data.usage,
-            &data.choices,
-            |choice| CompatibleChoiceData {
-                finish_reason: if choice.finish_reason == Some(FinishReason::ToolCalls) {
-                    CompatibleFinishReason::ToolCalls
-                } else {
-                    CompatibleFinishReason::Other
+        Ok(Some(
+            openai_chat_completions_compatible::normalize_first_choice_chunk(
+                Some(data.id),
+                Some(data.model),
+                data.usage,
+                &data.choices,
+                |choice| CompatibleChoiceData {
+                    finish_reason: if choice.finish_reason == Some(FinishReason::ToolCalls) {
+                        CompatibleFinishReason::ToolCalls
+                    } else {
+                        CompatibleFinishReason::Other
+                    },
+                    text: choice.delta.content.clone(),
+                    reasoning: choice.delta.reasoning.clone(),
+                    tool_calls: openai_chat_completions_compatible::tool_call_chunks(
+                        &choice.delta.tool_calls,
+                    ),
+                    details: choice.delta.reasoning_details.clone(),
                 },
-                text: choice.delta.content.clone(),
-                reasoning: choice.delta.reasoning.clone(),
-                tool_calls: chat_compatible::tool_call_chunks(&choice.delta.tool_calls),
-                details: choice.delta.reasoning_details.clone(),
-            },
-        )))
+            ),
+        ))
     }
 
     fn build_final_response(&self, usage: Self::Usage) -> Self::FinalResponse {
@@ -250,7 +254,7 @@ pub async fn send_compatible_streaming_request<T>(
 where
     T: HttpClientExt + Clone + 'static,
 {
-    chat_compatible::send_compatible_streaming_request(
+    openai_chat_completions_compatible::send_compatible_streaming_request(
         http_client,
         req,
         OpenRouterCompatibleProfile,
@@ -262,7 +266,7 @@ where
 mod tests {
     use super::*;
     use crate::http_client::mock::MockStreamingClient;
-    use crate::providers::internal::chat_compatible::test_support::sse_bytes_from_data_lines;
+    use crate::providers::internal::openai_chat_completions_compatible::test_support::sse_bytes_from_data_lines;
     use crate::streaming::StreamedAssistantContent;
     use futures::StreamExt;
     use serde_json::json;
